Move createrepository from module to service layer
